Move pants.init.subprocess into pants.process
ProcessManager, NailgunTask and GlobalOptions BUILD files
had no declared dependency on the imported module
pants.init.subprocess.

This led to runtime failures during test runs that looked like

```

                  .pants.d/pyprep/sources/ae8f7a03d241798e1604e92b254c9c09a614234a/pants/pantsd/process_manager.py:19: in <module>
                         from pants.init.subprocess import Subprocess
                     E   ImportError: No module named init.subprocess
```

Adding the BUILD dep revealed a cycle:

```
Exception message: Build graph construction failed: ExecutionError 1 Exception encountered:
Computing Select(Specs(dependencies<=tuple>=(SingleAddress(directory=u'tests/python/pants_test/java/distribution', name=u'distribution'),), matcher<=SpecsMatcher>=SpecsMatcher(tags<=tuple>=(), exclude_patterns<=tuple>=())), =TransitiveHydratedTargets)
  Computing Task(transitive_hydrated_targets, Specs(dependencies<=tuple>=(SingleAddress(directory=u'tests/python/pants_test/java/distribution', name=u'distribution'),), matcher<=SpecsMatcher>=SpecsMatcher(tags<=tuple>=(), exclude_patterns<=tuple>=())), =TransitiveHydratedTargets, true)
    Computing Task(transitive_hydrated_target, tests/python/pants_test/java/distribution:distribution, =TransitiveHydratedTarget, true)
      Computing Task(transitive_hydrated_target, src/python/pants/java/distribution:distribution, =TransitiveHydratedTarget, true)
        Computing Task(transitive_hydrated_target, src/python/pants/java:util, =TransitiveHydratedTarget, true)
          Computing Task(transitive_hydrated_target, src/python/pants/java:nailgun_executor, =TransitiveHydratedTarget, true)
            Computing Task(transitive_hydrated_target, src/python/pants/pantsd:process_manager, =TransitiveHydratedTarget, true)
              Computing Task(transitive_hydrated_target, src/python/pants/init/process:process, =TransitiveHydratedTarget, true)
                Computing Task(transitive_hydrated_target, src/python/pants/core_tasks:core_tasks, =TransitiveHydratedTarget, true)
                  Computing Task(transitive_hydrated_target, src/python/pants/pantsd:process_manager, =TransitiveHydratedTarget, true)
                    Throw(Dep graph contained a cycle.)
                      Traceback (no traceback):
                        <pants native internals>
                      Exception: Dep graph contained a cycle.
```

The solution taken here was to move the subprocess thin wrapper
into the prexisting pants.process module. This makes the diff
of this patch less intuitive, because those modules already
brought that in. But the next-best place would have maybe been
next to the process_manager, which would require the jvm
backend to depend on pantsd and all that nonsense.

This seems like the best fit with the lightest cost - feel free
to move elsewhere as you see fit in a followup.
