+++ src/python/pants/pantsd/process_manager.py
@@ -16,7 +16,7 @@ from contextlib import contextmanager
 import psutil
 
 from pants.base.build_environment import get_buildroot
-from pants.init.subprocess import Subprocess
+from pants.process.subprocess import Subprocess
 from pants.process.lock import OwnerPrintingInterProcessFileLock
 from pants.util.dirutil import read_file, rm_rf, safe_file_dump, safe_mkdir
 from pants.util.memo import memoized_property
