+++ src/python/pants/process/BUILD
@@ -6,6 +6,7 @@ python_library(
     '3rdparty/python:future',
     '3rdparty/python:fasteners',
     '3rdparty/python:psutil',
+    'src/python/pants/subsystem',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:process_handler',
   ]
